+++ src/routes/LandingRoutes.js
@@ -1,24 +1,9 @@
 import React from 'react';
 import { Route } from 'react-router';
-import { Link } from 'react-router-dom';
 import { withRouter } from 'react-router-dom';
+import Landing from '../containers/Landing';
 import SingIn from '../containers/SingIn';
-import Button from '../components/Button';
 
-const Landing = () => {
-  return (
-    <>
-      <h1>Landing</h1>
-
-      <Link to='/signin'>
-        <Button buttonType={'primary'}>Sign In</Button>
-      </Link>
-      <Link to='/signup'>
-        <Button buttonType={'primary'}>Sign Up</Button>
-      </Link>
-    </>
-  );
-};
 const Register = () => <h1>Register</h1>;
 const LandingRoutes = () => (
   <>
